Refactor header into a reusable component
